+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 - Initial Release
+
+## 0.0.2
+
+- Fix initial camera front
+- Improve switch camera lens